Добавляет абстрактный класс
